disable automatic index creation
as of recently, the data ingestion container sometimes fails at the
attempt to create elasticsearch containers, because the simultaneously
started backend app container is in the process of creating those at the
same time. This prevents application context to properly initialize.
We should be able to fix this by just creating the entity indices
ourselves at application startup, and handle the expected exception
silently.